requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymupdf==1.24.10
pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
groq==0.11.0
pandas==2.2.2
sentence-transformers==3.1.1
numpy==1.26.4